split file function.php to (function.php, database.php, download.php).
